backend/app/db.py
import os
from dotenv import load_dotenv
from motor.motor_asyncio import AsyncIOMotorClient

load_dotenv(".env")
MONGO_URI    = os.getenv("MONGO_URI", "mongodb://localhost:27017")
mongo_client = AsyncIOMotorClient(MONGO_URI)
db           = mongo_client["tdp_prototype"]

__all__ = ["db", "mongo_client"]
