backend/requirements.txt
fastapi
fastapi-crudrouter
uvicorn[standard]
motor
pydantic
python-jose
python-dotenv
pydantic[email]
lmstudio
openai
networkx>=3.2
pypdf>=3.9
python-multipart
google-generativeai